+++ config/routes.rb
@@ -79,11 +79,10 @@ Hybrid::Application.routes.draw do
     end
 
     resources :joins
+    resources :rules
   end
 
   resources :source_file_types
-
-  resources :source_rules
 
   devise_for :users, controllers: { registrations: 'contour/registrations', sessions: 'contour/sessions', passwords: 'contour/passwords', confirmations: 'contour/confirmations', unlocks: 'contour/unlocks' }, path_names: { sign_up: 'register', sign_in: 'login' }