+++ db/migrate/20130625191017_rename_source_rule_to_rule.rb
@@ -1,0 +1,5 @@
+class RenameSourceRuleToRule < ActiveRecord::Migration
+  def change
+    rename_table :source_rules, :rules
+  end
+end